Move test where it belongs.
